# colexecjoin: optimize building output on the left in cross joiner